Support multiple STL files across same module
* Support multiple STL files across same module

Modules can be split across multiple files. Example usage is reflected
in example/example.test, which imports example_base.stl (event
definitions, transition definitions, etc.) and example.stl (transition
instances).

* Allow STL constants to be defined in manifest

Constants can be declared in the STL without a value, and then defined
in the manifest via the 'constants' dictionary entry. Constants are
referenced in the manifest by '<module>::<constant name>'.

* Pass --manifest-args as a list instead of a string

If a manifest variable was a string with a space in it (for example,
--manifest-args="the_string='this has spaces in it'"), it could not be
parsed since we were splitting on whitespace. Now, we just pass the args
as a list (e.g. --manifest-args "this_string='has spaces'" "foo='bar'").

* Fixed quotes in README.md

Example code and commands should be copy/pastable directly from the
README.

* Change Uniqueness qualifiers to not require previous value

This is a simplification of the built-in UniqueInt and UniqueString
qualifiers.